Add a critical error if DB connections are shared across the fork
rt-server takes pains to close down the database connection before
handing off control to the PSGI server, such that the database
connection is not preserved across the `fork`.  DB connections which
are so shared result in undefined behavior -- postgres, for instance,
will reuse the same statement ids on the handle in different
processes, resulting in errors of the form:

    DBD::Pg::st execute failed: ERROR:  prepared statement "dbdpg_p4068_1" already exists

However, nothing prevents extensions from recreating the DB connection
during their `to_app` methods -- RT::Extension::REST2-1.00 erroneously
does so, for instance.

While we could force-unset the connections, this would serve as a
continuation of the mostly-silent error.  Instead, add an explicit
error which causes server startup to fail, encouraging users to check
for updated versions of plugins which will right the error.