Fix bug on groups granted 'SeeGroupDashboard' or 'ShowSavedSearches'
Groups granted 'SeeGroupDashboard' or 'ShowSavedSearches' rights
on a second group, previously, could only see the dashboard or
saved searches if the group members were also in the second group.
This fix allows for members of one group with these rights on
another to view the dashboard or saved searches of the other group
without membership.