Ignore externalized data when serializing.
Adding an option to rt-serializer that lets you leave externalized data
externalized when serializing. The default behaviour is to pull all
externalized data in to the serialized data. If you have lots of
externalized data this could make a huge difference when serializing and
possibly importing the data after that.