Add option to silence 'Change of config' warnings
When an RTIR configuration option is updated in RT_SiteConfig.pm,
a warning message is added to the logs stating the update will
overwrite its default config option. The behavior the message is
expressing is expected if the change was intentional. This warning
message might not be very helpful to some users and it creates
additional noise in the logs. This warning message has been moved
to the debug level to allow users the option to view these messages
or not by enabling or disabling 'debug' with the RT config
$LogToSTDERR.