Silence "numeric" warnings in GD::Graph for stacked bar charts
We want to show extra info such as status or queue names in bar values,
but values are supposed to be numeric. Though strings like "2 open"
could be automatically interpreted to numbers, perl warns "isn't
numeric" in this case.

The other related warning is "Use of uninitialized value $_ in numeric",
which happens if there are absent stacks(i.e. 0 tickets in some groups).