Hide Queue if Requestor does not have right to see
It is possible for a User to be a Requestor on a Ticket and not have the
right to view the Tickets Queue. If this is the case, when editing
basics the current Queue cannot be displayed as the dropdown value for
Queue. Thus resulting in the next available Queue being submitted as the
new Queue for the Ticket. This causes unwanted Queue changed to the
Ticket.