remove group member limit for shared setting access(dashboards/savedsearches)
Previously only members of a group can access the group's dashboards and
savedsearches. This limit is inconvenient and also unnecessary
considering we have corresponding rights to control the access limit.

The new behavior drops member limit and totally depends on Rights. E.g.
as long as a user has "ShowSavedSearches" right on the group, it's
totally fine to allow the user to access the group's savedsearches.

Global "ShowSavedSearches" means permission to access all the group
savedsearches.