warn if it's permission that prevents site configs from being loaded.
For core configs like RT_Config.pm, RT dies with a long message if there
is a permission issue. But for *optional* site configs, RT simply
ignores if they can't be loaded, because usually it means the site
configs don't exist.

Since perl 5.18, the "Can't locate ..." message has been improved so we
can tell if there is a permission issue. Thus, we can easily warn people
if a permission issue is found, since it's most likely a configuration
issue.