Fix the issue that invalid queries to FromSQL could pass validation
An invalid query like:

    Status = 'open' and LastUpdated < yesterday

could pass validation and be wrongly parsed as:

    SELECT main.* FROM Tickets main WHERE (main.IsMerged IS NULL) AND (main.Type = 'ticket') AND (main.Status = 'open')

Since b68c84f0(Switch to parsing into a parse tree as an IR),
RT::Tickets::_parser uses RT::Interface::Web::QueryBuilder::Tree to
validate and parse queries, which doesn't "die" but returns errors if
there are any found.

_parser should directly "die" if errors are found in ::Tree->ParseSQL,
especially that RT::Tickets::FromSQL relies on this "die" behavior to
determine if the query is valid or not.