Always enable recipient checkboxes
The "disable" behavior is introduced in f0a7105d and 4b11cf37, which is
mainly to indicate that recipients checkboxes are being updated and
shouldn't be changed. It had a js execute order issue, which was fixed
in db6ef531. But from real user reports, there is still possibility that
checkboxes are not enabled in time, which causes RT to not send
notifications.

It introduces so much mess but gains so less, not mentioning that the
0.3 opacity is already an indicator of "watch out, the section is being
updated".