Switch to Clone::clone to copy config structures in Obfuscate callbacks
Storable::dclone doesn't support CODE and REGEXP, which are commonly
used in configs, so you may encounter errors like "Can't store CODE
items" or "Can't store REGEXP items" on system configuration page.

There is a quick workaround to make it support CODE, but seems there
isn't a simple way to support REGEXP. Migrating to Clone::clone could
fix both.

Note that Clone is not a new dependency because DBIx::SearchBuilder
requires it for decades, adding it to rt-test-dependencies is just for
consistency since now we require it directly in RT.

Fixes: I#33019, I#33581