Hide texts in multipart/related only if it has a perferred alternative
In 4a38585f, we added code to handle the following structure:

    multipart/alternative
        text/plain
        multipart/related
            text/html
            image/png

That commit treats "multipart/related" as "multipart/alternative",
assuming "multipart/related" is always under "multiplart/alternative",
which is sadly not always correct, recently we encountered an email with
a structure like:

    multipart/related
        text/plain
        image/png

It doesn't have a corresponding "text/html" alternative, so if user
prefers rich text, only the "image/png" shows up on ticket display page.
This commit fixes it by checking if "multipart/related" is really under
"multiplart/alternative".