Store log messages until RT::Logger is initialized
Log messages generated during startup configuration processing
can't use RT::Logger because it hasn't been initialized yet.
Store them and output the messages as soon as the logger is
available. This allows the log messages to be output in the context
of the configured log level, so debug messages won't be output
if RT is configured to show only warn or error.