Filter queue custom roles by checking current user's right
Previously, "Roles" rerturned all the roles registered for the object
and didn't consider if current user has right to see them or not.
"Role" and "HasRole" have the same issue.

This commit adds the rights check to AppliesToObjectPredicate, which is
called in "Roles".  As "HasRole" calls "Roles" internally, it's fixed
automatically.  "Role" is tweaked to add the rights check via "HasRole".