Refactor ticket create and update to use new SelectArticle
Remove multiple article selection options from the ticket update
page and replace with a single article selection control. This
new article selector is a dropdown like the former hotlist,
but converts to an autocomplete when the number of items in the list
exceeds the DropdownLimit configuration.

This change removes some extra code that worked to load a ticket
object in IncludeArticle. The template no longer needs it, but
it was also passed to a callback. Extensions that use this
callback may need to make changes to handle receiving an empty
ticket object.