Pass UTF-8 decoded data to Create method for rt-importer on Pg
MySQL and Pg are different: serializer generates UTF-8 encoded bytes for
MySQL and UTF-8 decoded string for Pg, respectively. This is not a
problem if you stick to the same database, but when you try to migrate
an RT db from MySQL to Pg, encoding issues could happen because Pg
expects UTF-8 decoded values instead of encoded ones. MySQL on the other
hand, isn't that picky.

This commit fixes this particular issue by passing UTF-8 decoded string
to Create on Pg. But for cases where the data is not valid UTF-8(quite
rare), original values will be stored.