Print raw transaction info on missing queues
This change adds printing of OldValue and NewValue for Queue changes when no Queue is found.

One use case occurs on partial migration of a database, where some queues are not migrated. The transactions for the changes in the ticket history will only print # for queues not migrated, while this change will print the actual value of OldValue and/or NewValue.