Don't enable/disable related groups when enabling/disabling custom roles
This is for performance: as each ticket has a correponding group for a
custom role, it could be time consuming when there are many tickets
involved.

Since groups are not in sync now, we need to exclude disabled/unapplied
custom roles when listing/checking rights.

This also fixes a bug that when you move a ticket from queue A to B,
people in a custom role that is applied to A can still get rights via
the custom role in B, even if the custom role is not applied to B.