Use user timezone for date "=" queries in ticket search
Take query "Updated = '2018-10-17'" for example, it means tickets
updated on 2018-10-17, which should always be equal to query:

    Updated >= "2018-10-17 00:00:00" and Updated < "2018-10-18 00:00:00"

Previously, if current user's timezone is not the same as the default
one(in RT config), e.g. if user and default are "UTC" and "-04:00",
respectively, the above "Updated = '2018-10-17'" query is actually equal
to:

    Updated >= "2018-10-16 20:00:00" and Updated < "2018-10-17 20:00:00"

which is not right.

This commit fixes this inconsistence.