Load Content rather than Attribute object for RecentlyViewedTickets
Previously, if the RecentlyViewedTickets content somehow was set
to undef, $content still evaluated to true because it was an
RT::Attribute object. Calling the Content method then resulted in:

    Can't use an undefined value as an ARRAY reference

This caused a 500 error for the end-user.

Loading the content provides the intended true/false value and handles
the empty content case.