Make RescueOutlook handle Office HTML mails better
Some Office-generated HTML mail will have extra newlines added the same way
Outlook adds them, and we should clean these up too. The lines are slightly
different than in the Outlook case, so we do three substitutions, one for each
case seen in the wild. If any of them changes the mail body, we'll save the
cleaned up version of the mail.

Added a test case based on an Office-generated mail found in the wild.