Skip merged tickets in role groups validation
Those role groups are not actually used any more, so there is no need to
re-create them.

Even worse, for "Owner" role group, a simple re-creation is not good
enough as the group always expects a user in it, which normally could be
covered by other "resolve" code, but sadly not for merged tickets. So if
we create an "Owner" role group without a user in it for a merged ticket
here, rt-validator will always warn something like:

    The owner of ticket #2 is inconsistent. The denormalized owner is user #6
    but there is no normalized owner.
    Found a ticket that has no normalized Owner group member.

Let's just not create those useless role groups.