bin/rt should print password prompt to STDERR
Ticket #34216: Feature request: bin/rt should print password prompt
to STDERR

I've been using bin/rt in shell scripts, redirecting STDOUT to a file,
or piping it to other processes. This will hang, waiting for a password
prompt, if I haven't previously logged into rt.

Printing the password prompt to STDERR rather than STDOUT will send the
password prompt to the screen, even if STDOUT is redirected, and also
has the advantage of not polluting STDOUT with the password prompt.

This patch will cause the password prompt (as well as the 'Press CTRL-C
if you do not want to continue' warning) to print to STDERR rather
than STDOUT.