Refactor ReplaceAttachments method mainly to cover more attachments
Previously we only replaced attachments of "text/plain" and "text/html",
which is good enough for "Content" replacement but sadly not for
"Headers", as we missed quite common types like "multipart/mixed" and
more. This commit imporves "Headers" replacement by covering all the
types of "multiplart/*" and also embeded email ones like
"message/rfc822", 'message/partial' and 'message/external-body'.

Code is also refactored to create Munge txns as soon as possible for
better consistency, and the memory usage is also reduced as we don't
need to keep all the ticket objects.