Make ReplaceHeaders/ReplaceContent return true only if replacement happens
We filter attachments in RT::Attachments::ReplaceAttachments by the
search string in advance, so replacement is supposed to always happen in
the inner calls of ReplaceHeaders/ReplaceContent, which makes the
statement of

        return ( 1, $self->loc('Headers cleared') );

always be called and the statement of

        return ( 1, $self->loc('No content matches found') );

never be called.

Logically both statements above are not quite correct considering
attachments that replacement methods are called on might not have
matches.

This commit fixes this, so we can always use the return values to
indicate if there are replacements or not.