Pass ticket object using Object parameter rather than TicketObj
ProcessObjectCustomFieldUpdates accepts an object to act on via
the generic parameter 'Object'. However, for tickets, it is often
called in a sequence with several other calls that pass the
available ticket object with TicketObj rather than Object.

The function loads a fresh object using an id found in other
parameters, so it works even when the Object parameter is missing.
However, when running under DryRun, the newly created and loaded
ticket object doesn't have the DryRun value set. In the case
of PreviewScrips, this causes the commit phase of scrips to run
when a custom field argument was passed and processed.