Avoid "Wide character in print" warnings when generating shredder SQL output
Depending on the database type, RT::Record::_AsInsertQuery could be utf8
decoded characters(on Pg and Oralce) or utf8 encoded bytes(on MySQL and
SQLite).  When encountering wide characters(on Pg and Oracle), perl
issues the "Wide character" warnings as it expects bytes.

This fix uses the same code as in RT::Record::__Value, for consistency.