Defer AJAX recipients update a little bit to get form's latest status
We fire updateScrips on input changes, including UpdateType, Status,
Owner, UpdateCc, UpdateBcc, UpdateContent, etc. Most text input changes
are triggered when we focus on other elements of the page, and things
become tricky if the other elemnts are recipient checkboxes, e.g.

    1. Add some text to the message body(the plain textarea)
    2. Check or uncheck one of the checkboxes in recipient list

The 2nd step not only changes the checkbox's status, but as the focus
changes from textarea to checkbox, updateScrips is also fired.

The bad thing is the focus change happens right before the checkbox
status change, thus updateScrips is fired a little bit earlier. When we
serialize inputs in updateScrips, we can only get the old status of the
checkbox.

This commit fixes this issue by deferring the serialization a little bit
to get the latest status of the form.