Switch to "POST" when updating ticket search result format via AJAX
Adding all the form fields to URL could exceed web server/browser URL
length limit, so move it to request body instead.

This commit is to fix the issue of "Failed to update format. Reason:
error" because of the over-sized URL when updating section "Display
Columns" in ticket search builder.