Accept a ShowHistory arg as an override for the default setting
This allows a different loading mode for the history for
an single ticket display. This is useful for debugging when
the source HTML is needed for transactions and the default
scroll mode shows only "Loading...". It is also useful for
the TicketToPDF extension which needs to load the entire history
to generate a PDF.