Don't recursively add members to ticket role groups in CachedGroupMembers
CachedGroupMembers is used to quickly check if a user/group is a member
of another group, recursively. The table contains flat relationships
between groups and their members(including recurisve members), e.g.

    - top
      - sub
        - user

In CachedGroupMembers, the rows are like(simplified and also excluded
self to self rows):

    Group Member
     top   sub
     sub   user
     top   user

You can see there is a direct relationship between "top" to "user" in
CachedGroupMembers.

The performance becomes quite bad when processing updates of groups that
are in roles of a huge list of tickets. e.g.

    - AdminCc(for ticket 1)
      - group1

    - AdminCc(for ticket 2)
      - group1

    - AdminCc(for ticket 3)
      - group1

    ...

If we add a new member to "group1", we need to create "AdminCc <-> the
new member" rows in CachedGroupMembers for all the tickets above, which
could be quite time consuming: as the complexity is O(n), it will be
linearly slower as the ticket list grows.

To get around it, let's not create those rows in CachedGroupMembers and
tweak logic in RT instead.