Check direct group members in recursive member methods for ticket role groups
As we don't maintain a full flat relationship in CachedGroupMembers for
ticket role groups(in 9444ce14), we need to check all their direct group
members too, which have full flat relationships themselves.

Note Domain is set for all the rows, the expression of "$self->Domain //
''" is to eliminate uninitialized warnings for cases where $self hasn't
loaded a row yet(it happens in tests).