Reset time part after date input is cleared
Once a date input is cleared, the datetimepicker UI reset the date part,
but sadly not the time part.  This commit resets the time part too, via
an official callback(beforeShow) so we don't need to touch original js.

Once the js lib fixes this issue, we can revert this commit.