Concatenate strings before encoding conversion to avoid character breakage
When a field value is composed by a list of encoded strings like
"=?UTF-8?B?...?=", converting each part one by one might break the
string, here is an example:

    Subject: =?UTF-8?B?5L2g5Q==?==?UTF-8?B?pb0=?=

The subject is consist of 2 Base64+UTF-8 encoded parts, of which base64
decoded content is "e4bda0e5" and "a5bd" in bytes, respectively. If we
convert them separately, since "e4bda0" is Chinese character "你", there
is no problem of that sub-part, while "e5" is actually incomplete, which
needs "a5bd" in the following part to represent another Chinese
character "好", the string is broken, i.e.

If we convert each part separately, we get bytes(rendered as "你���"):

    e4bd a0ef bfbd efbf bdef bfbd

If we convert it together, we get correct bytes(rendered as "你好"):

    e4bd a0e5 a5bd

This commit fixes this issue by concatenating succssive parts before
conversion.