Refine TODO test for content on merged tickets
Setting ShowSearchResultCount on by default puts the string
"(Found n ticket)" next to searches on the home page. The search
in this test also looks for that string, but in the context of
a search result. Having it appear on the home page caused the
TODO test to incorrectly pass.

Refine the search to match only on the string when found in an
h1 header line.