Switch to bootstrap card for titleboxes
As old themes stick to old rollup implementation, we need a convenient
way in js to distinguish new themes from old ones. For now, we check if
jQuery(e).collapse is defined, and to make it really work, we need to
exclude bootstrap js library from old themes, which will be in a
separate commit.