For saved searches, split objects lists for creating and loading
Different rights manage creating/editing saved searches and
loading them, so split loading the list of valid objects into
two different data structures.

Also use the corresponding wrapper methods to load the objects
rather than calling the _PrivacyObjects method directly.

This change fixes a bug in the Privacy menu. Previously it
loaded all groups the user is a member of without checking the
EditSavedSearches right. If the user then tried to save a search
for a group without that right, they would receive an error and
the search would not be saved. The dropdown now checks for this
right and only shows group options that the user can save to based
on rights.