Clarify external auth logging when users are not found
Add a debug message to log when an account could not be found
in the configured external sources.

Create can be called in different contexts, some providing Name and
some providing EmailAddress, for example. In these cases, one or
the other may not have a value and this is not necessarily a
misconfiguration. Clarify the message to avoid giving this
potentially misleading information.