Automatically load more txns to fill browser window on scroll history mode
For various reasons(e.g. ticket metadata section is collpased or some
txns are skipped), it's possbile that ticket display page only shows a
couple of txns by default, leaving space at the bottom of the page
without scroll bar, which could mislead people to think that all the
txns are already loaded.

This commit fixes this issue by automatically loading more txns when
there is still space left in the browser window.