Refactor custom date range code to make it easier to expand
* _ParseCustomDateRangeSpec

Previously we returned a 5-elements list, which was too many and it's
inconvenient to add more config items to the spec if we want to return
them from _ParseCustomDateRangeSpec.

Now it returns a hash instead.

* $op

As only '-' is supported and it's probably also the only sensible
operator for 2 dates, there is no need to extract it.