Support role searches like Owner = CF.cid or Owner = Creator
Previously RoleLimit did't consider the value could be a column name, so
it always tried to load the user using the value as Name/EmailAddress.

This commit skips loading user for columns(not quoted values) and sets
field to compare to id as usually we store id in columns like Owner,
Creator, etc.