Reduce code duplication of checking formats of CustomFieldGroupings
Various formats are supported in the structure of CustomFieldGroupings
and we canonilize them in PostLoadCheck. Since the PostLoadCheck will
always be run in real world, it's fine to assume the config is well
canonilized in code.

Current duplicated code is not much, but as we are going to support more
formats there, the duplication will be a pain to maintain.