Simplify SQL chart style output
The new added structure split the content into at least 2 rows(label
itself took one row), which is not necessary. The original version is
generally better.

Note that div.query is in div.chart-wrapper in original version.  This
commit reverts back to it and reuses the same style in old themes.