Log number of records returned from LDAP search
ExternalAuth requires a single result from an LDAP search
to use the returned data to populate a user record. Since it
will stop processing and move on to other configured auth
sources, log the number of entries returned to make it clear
why that user was skipped even though records were found.