Confirm record is defined before calling id
In cases where the $current_values iterator would run out of values
and return undef, the id call on $extra_item would cause an
"undefined value" exception. When called by a scrip, this causes
the scrip to die. It also left a dangling transaction when the Rollback
was not called, and this could cause the outer ticket transaction
to not commit properly.