Fall through to false on watcher rights check
_HasModifyWatcherRight previously checked known rights cases
and returned false when the user didn't have rights for the
corresponding change. If nothing matched, the fall through return
value was true, which meant that any watcher change that didn't
match was allowed. All new custom roles won't match the existing
watcher type checks, so users could by default edit custom roles
even without ModifyTicket or ModifyQueueWatchers.

Reverse the logic to return true for know rights and types and
default to 0 for any unknown cases. At this point, ModifyTicket
is required to modify custom roles on tickets and
ModifyQueueWatchers on queues.