Horizontally align label/input in forms again
padding-top of .form-control in our bootstrap version was set to 6px,
and as the input fonts are smaller, setting padding-top of .label to 5px
looks better.

For .label out of forms, this padding-top rule is removed as we want to
make the display layout denser(like "The Basics" widget on ticket
display page).