Allow login with all fields from attr_match_list
This modification allows the UserExists method to search the LDAP repository for any match of the fields specified with attr_match_list.
After a match is found (e.g. email address) the username is updated to reflect the value of the Name field.
The user can now use the value of any of the fields in attr_match_list to logon as long a the provided login data results in a unique match in the LDAP repositroy.