Clearly separate last asset and the following "Add an asset"
We intentionally disable the bottom border of the last item in
accordion, which looks fine mostly, until there are some other stuff
right below it(like the "Add an asset").

I chose to set top border for "Add an asset" instead of enabling bottom
border of the last asset item mainly because "Add an asset" doesn't
always exist(depending on if current user has ModifyTcket right)

The padding is to be consistent with previous accordion.