Remember collapse status of accordion items in ShowRequestor and ShowAssets
The logic is nearly identical to loadTitleBoxStates, so I directly
updated the function, renamed it and moved it out of titlebox-state.js.

Superfish menu in asset's accordion title will be broken if we show
accordion before initializing the menu. Thus I put related js to the
same jQuery block to assure the execute order.

I also added prefix "accordion-" to related HTML ids in case of name
collision in the future.