Enable bootstrap tooltips
For the tooltip event, bootstrap adds a div with a unique id and
modifies the svg with an aria-describedby attribute referencing that
id.  When the user leaves the tooltip event, the div is removed as
well as the attribute.  Since we're replacing the bookmark link and
image when the user toggles the bookmark, the aria-describedby
attribute is removed but bootstrap's div is left over, no longer
associated with any svg.

To ensure the tooltips are cycled correctly, we run the tooltip
hide then dispose methods for any tooltip divs before replacing
the bookmark links.