Set QUOTEVALUE only if necessary to use default behavior as much as possible
Previously QUOTEVALUE was set to 1/0 totally depending on if the value
is quoted or not in ticket SQL. This is incorrect and could cause SQL
generated by queries like "CF.foo LIKE 3" not been quoted correctly:

    ObjectCustomFieldValues_1.Content LIKE %3%

This issue was introduced when we supported columns as values. To fix
this issue, besides actually quoted values where QUOTEVALUE should be 1,
we only need to specifically set QUOTEVALUE to 0 when values are columns
and let the default behavior handle other cases.