Migrate charset to utf8mb4 in mysql
This allows 4-byte characters, such as emojis, in fields.
This feature is supported properly in mysql 5.7.7 and later.

It requires the database to be converted to utf8mb4 to
support those characters.

MySQL version should be >= 5.7.7
MariaDB version should be >= 10.2.5

https://dev.mysql.com/doc/relnotes/mysql/5.7/en/news-5-7-7.html#mysqld-5-7-7-feature
https://mariadb.com/kb/en/changes-improvements-in-mariadb-102/
https://mariadb.com/kb/en/mariadb-1025-changelog/ (search for utf8)

Before this version utf8mb4 tables could not have indexes with type VARCHAR(255):
the default size for index entries was 767 bytes, which is enough for 255 chars
stored as at most 3 chars (the utf8 format), but not as 4 bytes (utf8mb4).
5.7.7 sets the default index size to 3072, so all of RT indexes are now OK.

Detection of MariaDB is done on the version string (it must match "mariadb"),
which is the case in both Debian and RedHat based distributions, and likely
everywhere else.

The innodb_log_file_size check is removed as it's not an issue any more in
5.7, see also https://dev.mysql.com/doc/relnotes/mysql/5.6/en/news-5-6-20.html