Add Create Ticket link to Home menu
On most pages, the Create Ticket button is available in the title bar.
However, when working on Assets or Articles, the Create button changes
to Asset or Article leaving no obvious way to create a ticket.

Add a Create Ticket link that is available anywhere in RT. Assets
and Articles have their own menus with a similar Create link. Use
the Home menu since RT doesn't have a stand-alone Ticket menu.