Refactor SectionMap to get the whole structure from RT_Config.pm
Previously we had pre-defined tabs/sections in RT::Config, and the
structure is not consistent with RT_Config.pm, people might get
confused when they check the doc of RT_Config.pm and web UI.

This commit refactored the parse logic of RT_Config.pm to use it as the
single source of the whole structure.

The structure of RT_Config.pm is supposed to be:

=head1 Tab
=head2 Section
=head3 Subsection
=item Option

We will update RT_Config.pm in another commit.

I also update the "missing META" log level from "debug" to "warning", to
catch the missing items a bit more easily.