Resolve the inconsistent $Class param in Search and CollectionAsTable
$Class on search pages is supposed to be "RT::Tickets" or
"RT::Transactions", but is "RT__Ticket" or "RT__Transaction" in
CollectionAsTable.

As we support not just tickets but also transactions on search pages, we
need to pass $Class from CollectionAsTable/Header back to search
pages to make links like "ID", Subject", etc. work.

Considering CollectionAsTable actually needs ColumnMapClass and it's
easy to get that from collection class, here we switch to
"RT::Tickets"/"RT::Transactions" in CollectionAsTable, so we don't need
to do conversions.