Add styling for action buttons
This commit updates the action buttons to be similar to the new
icons in the elevator themes.

The history css contained a number of margin definitions that were
cascading to the action buttons, making the new styling quite
complicated to adjust.  This commit removes a number of margin
definitions to simplify control, as well as tighten the general
spacing of the elements to make each section more compact.