Dynamically adjust width for subject and page menu
The goal is to use available space as much as we can. The tricky thing
is both subject and page menu could be too wide. To handle cases like
that, first we let subject take up space it needs, and then let menu
take up left space. If the left space is too small, force subject/menu
width to 0.6/0.4 layout.

Note that we need subject to be "inline-block" so .width() could return
the real space it takes.