Support arbitrary height of #header to show long wrapped subject
This commit doesn't fix the possible text overlap of subject and page
menu, it mainly supplies a good base for it.

To not hardcode height of "#header", we couldn't hardcode its parent
"#rt-header-container" either, to keep the box shadow effect of the
first row("#app-nav", "#topactions", etc.), corresponding setting is
moved to "#quickbar".

For short subjects, the UI is the same as before.