Fix topactions menu overlap
This commit fixes the topactions menu overlap by calculating the
available space for app-nav based on the width of topactions.

This commit also fixes the topactions input animation by only
easing if there is enough space available.  The easing definitions
in the CSS are now bound to the "expand" class, which is added and
removed by the JS in BeforeNav after calculation.

The calculation for the animation space takes into account the
topactions width, topLevelMenu width, logo width, and 80px (the max
difference being 6em plus a little extra).