Update DashboardSelectionUI to elevator-light
Other than general design changes, this commit also fixes the reset
to defaults for the pages by using the reset logic from
Prefs/MyRT.html.  Additionally, it changes the post to use id as a
query param instead of form variable, so the url is correctly
including id after post.