Change layout of footer
This commit updates the footer by moving the version to the left,
removing the GPL language and logo, updating the Sales and Support
language on the right, and updating styles.

Once logged in, the version and copyright share the center middle.

Additionally, this change removes styling from the layout.css in
exchange for bootstrap defined classes.