Remove search string including numbers in ticket autocomplete search on select
Previously we only removed non-integers, which is not enough as the
serch string could be something like "RT 5 alpha", in which case "5"
wouldn't be removed.

Search string starts with non-integers and lasts till the end, so we can
remove all the items after the first non-integer.

Note that integer stringification("var str = term + '';") was only for
the new value. Since now we add the new value a bit later, there is no
need to do it any more.