Convert more OR'd clauses to use "IN" for performance
The search results should be the same, but the performance is better.
E.g. we found that "IN" version of user right check could be 4x faster
in some RT instances.

This is mainly to cover ticket display page. Another major OR thing left
on that page is Status clauses in TicketSQL, which will be handled in
another commit since the implimentation will be quite different.

See also b14df0ee4f