Revert "Add method for lifecycle rights check"
This reverts commit e4071599e6642f7ac35578bcc74652f5de047a8e.

The purpose of the context object param to Lifecycle method is to check
right on the passed ticket/asset object in case corresponding rights are
granted at role level. Unfortunately this is incomplete, as it can't fix
similar right issue on ticket/asset create.

As lifecycle is not a sensitive info generally, we will bypass right
check when necessary, which will be implemented in another commit.