Show string version of priorities to end users when possible
This is to obey the rule of least surprise. If PriorityAsString is
enabled, we should show priorities as string consistently.

E.g. when user updated Priority from "Low" to "Medium", the successful
message and corresponding transaction should show:

    Priority changed from 'Low' to 'Medium'

instead of:

    Priority changed from '0' to '50'

Another example: in search builder, when user selected "Priority" "less
than" "High", we should add

    Priority < 'High'

instead of:

    Priority < 100

But when configs conflict somehow like:

    General => { Low => 0, Medium => 50, High => 100 },
    Support => { Low => 20, Medium => 70, High => 200 },

Since there is no one-to-one map of String => Number, to avoid
confusion, we have to show priorities as numbers(e.g. in search builder
when invovled queues conflict). I believe it's a rare case.