Merge multiple schema changes of Attachments into one SQL
Attachments is a big table, and the performance is evidently better
using one SQL. Besides that, it also could reduce disk space usage for
databases that need to make table copies on schema changes.

With this commit, we can be sure that required free disk space won't
exceed the size of current database.