Refactor DashboardSelectionUI
To align this feature with the rest of RT, this commit removes the
AJAX call, moves the helper code into a subroutine, and updates
the pages to post to themselves.

Rather than post to the helper, the JS now writes the values to the
form as hidden inputs.

Additional information about each arg previously contained in the
JSON is now passed to the new subroutine to be used before updating
the preferences.