Correctly assign both return values of $Ticket->CurrentUserCanSetOwner() in html/Elements/Tabs and html/Ticket/Elements/ShowSummary.
This fixes a bug in the ticket menu tab for users without rights to set
the current ticket owner: the People tab will no longer be shown in that
case.

Similarly, the 'People' tab under 'Ticket Metadata' on the ticket
summary page will no longer be clickable in that case.

This bug was introduced in commit 43f0e872019ca3abd20168730905477dac98785d