Decode mailgate args
Need to decode mailgate args before they reach the database or else Postgres
can run into encoding issues.

Using a queue name in the mailgate command that has ascii characters
can result in RT not being able to find and load the queue if the queue
name string is not decoded.