Display the attachment name in ticket history
If a ticket has a lot of attachments it can be cumbersome to hover
over the download link to find the file you want. This commit makes
the filename available on the page and moves the filetype and size
into the tooltip instead.

This commit additionally corrects an alignment issue where the
downloadattachment div overlapped the above element's border if
there were many attachments in one transaction.