Update background and borders for elevator-dark
This commit sets a lighter background for the whole of the dark
theme as well as change background colors and borders for many of
the elements within.

Because we're initially setting colors using a "blanket" selector
for dark theme (.darkmode, .darkmode *), we have to manually
specify and fix each individual element accordingly. This means for
now using * within more sections to ensure child elements don't get
the color definition using the "blanket" selector.

Ideally, we should remove the use of the blanket definitions and
convert both elevator light and dark to use more selectors with
higher specificity. We can then get away from using "important" so
frequently to override definitions for each theme.