Fix ticket history transaction link colors
This is a copy of the colors from elevator-light. While not ideal
to replicate the definitions here, we need to specifically override
the cascading background color definition at the top of the
elevator-dark css page. Once we remove the cascading definition,
these transaction link colors would be much better to inherit from
elevator-light.