Fix missing chevrons in menus for IE11
IE11 is a lot more specific about what it allows for svg url
background definitions. Rather than change the svg we know works
for every other browser, load the chevron pngs for IE11 only.

Targeting IE11 in CSS is much easier and cleaner to do in JS. If
IE11, add the 'IE11' class to body to make for easier selector
specificity.