Refactor autocomplete for the Owner field in search results.
Remove ID from owner field, replace it with a class.

Move the JS function to util.js. The query parameter is passed as the
data-query attribute in the HTML. The name of the helper ("Owner"
in this case) is passed as the data-field attribute, so the function
can be re-used.