Migrate owner autocomplete to the general autocomplete
The initial reason is to remove the duplicate "Owner" id in search
results when InlineEdit is enabled.

Previous owner autocomplete had a simple client cache mechanism, but the
general autcomplete feature doesn't have that yet. We can come back to
this if there is a performance issue in the future.