Initial check-in of RT::Extension::REST2
PodViewer functionality from the extension is intentionally
omitted because all usage docs are in the main REST2.pm file
and as part of RT it will be published with the RT documentation
on the website.

Initial check-in of RT::Extension::REST2

PodViewer functionality from the extension is intentionally
omitted because all usage docs are in the main REST2.pm file
and as part of RT it will be published with the RT documentation
on the website.

remove Root.pm