Explicitly encode to UTF-8 for json serializer only when it's Pg/Oracle
Serializer generates UTF-8 decoded string when it's Pg/Oracle, which is
different from MySQL/SQLite.

We need to check database type instead of the utf8 flag of json encoded
output is because JSON doesn't care what data it gets and the utf8 flag
totally depends on the JSON object(i.e. whether "utf8" is called or not)

See also 80f172bba3